tag logic finished

+++ pkg/app/app.go
@@ -12,9 +12,9 @@ type Response struct {
 }
 
 type Pager struct {
-	Page      int `json:"page"`
-	PageSize  int `json:"page_size"`
-	TotalRows int `json:"total_rows"`
+	Page      int   `json:"page"`
+	PageSize  int   `json:"page_size"`
+	TotalRows int64 `json:"total_rows"`
 }
 
 func NewResponse(c *gin.Context) *Response {
@@ -30,7 +30,7 @@ func (r *Response) ToResponse(data interface{}) {
 }
 
 // request list and page infomation
-func (r *Response) ToResponseList(list interface{}, totalRows int) {
+func (r *Response) ToResponseList(list interface{}, totalRows int64) {
 	r.Ctx.JSON(http.StatusOK, gin.H{
 		"list": list,
 		"pager": Pager{
